+++ src/main/java/domain/Game.java
@@ -1,7 +1,9 @@
 package domain;
 
+import domain.position.DefaultUnitPosition;
 import domain.position.Position;
 import domain.position.Route;
+import domain.unit.Team;
 import domain.unit.Unit;
 import java.util.Arrays;
 import java.util.HashMap;
