+++ src/main/java/domain/Janggi.java
@@ -2,7 +2,9 @@ package domain;
 
 import domain.position.Position;
 import domain.position.Route;
+import domain.unit.Team;
 import domain.unit.Unit;
+import domain.unit.UnitType;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
