+++ src/main/java/domain/unit/Team.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.unit;
 
 public enum Team {
     HAN,
