+++ src/main/java/domain/unit/UnitType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.unit;
 
 public enum UnitType {
     GENERAL,
