+++ src/test/java/domain/position/PositionTest.java
@@ -1,10 +1,9 @@
-package domain.unit;
+package domain.position;
 
 import static domain.position.Position.INVALID_POSITION_EXCEPTION;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
